+++ include/utils/EDT.h
@@ -12,13 +12,13 @@ public:
 
   void computeDT(nav_msgs::msg::OccupancyGrid::SharedPtr grid);
 
-  double getDistanceAt(int idx) const;
+  double getDistanceAt(size_t idx) const;
 
 public:
   static constexpr double INF = std::numeric_limits<double>::infinity();
 
 private:
-  void computeDT1D(int start, int size, int stride);
+  void computeDT1D(size_t start, size_t size, size_t stride);
 
 private:
   std::vector<double> distance_transform_;
