+++ src/geometry/state.cpp
@@ -6,16 +6,16 @@ namespace geometry {
 
 State3d::State3d() {}
 
-State3d::State3d(int grid_x, int grid_y, int theta_bin)
+State3d::State3d(size_t grid_x, size_t grid_y, size_t theta_bin)
     : x(grid_x), y(grid_y), theta_bin(theta_bin) {}
 
 bool State3d::operator==(const State3d &other) const {
   return x == other.x && y == other.y && theta_bin == other.theta_bin;
 }
 
-State2d::State2d(int x, int y) : x(x), y(y) {}
+State2d::State2d(size_t x, size_t y) : x(x), y(y) {}
 
-State2d::State2d(const std::pair<int, int> &s) {
+State2d::State2d(const std::pair<size_t, size_t> &s) {
   x = s.first;
   y = s.second;
 }
@@ -27,10 +27,10 @@ State2d::State2d(const State3d &s) {
 
 std::size_t StateHash::operator()(const State3d &k) const {
   std::size_t seed = 0;
-  seed ^= std::hash<int>{}(k.x) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-  seed ^= std::hash<int>{}(k.y) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+  seed ^= std::hash<size_t>{}(k.x) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+  seed ^= std::hash<size_t>{}(k.y) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
   seed ^=
-      std::hash<int>{}(k.theta_bin) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+      std::hash<size_t>{}(k.theta_bin) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
   return seed;
 }
